requirements.txt
django==1.8.5
djangorestframework==3.3.0
gunicorn==19.3.0
